requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
psycopg2-binary==2.9.9
pyjwt==2.8.0
#datetime
uuid==1.30
argon2-cffi==23.1.0
uwsgi==2.0.26
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
marshmallow==3.21.2